backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
django-cors-headers==3.8.0
ipython
networkx==3.2.1